Cophonologies (#1)
Adds the ability to tag the input corpus with lexical categories, which are then parsed into separate cophonologies. It includes simulations on a hypothetical language , demonstrating the MDL learner's performance with and without lexical categories. It also introduces core functionality for initializing and running simulations, along with improved documentation and modularized methods for clarity. The project builds on the original taucompling/otml repository and extends its functionality.